Finalized irish spell library and created an irish spell file for testing
